% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mplot_for_Meta.R
\name{Mplot_for_Meta}
\alias{Mplot_for_Meta}
\title{Mplot for Meta data generated by METAL
MarkerName : eg. 1:10086:A/C or 1:10086:A:C
Usage eg. Mplot_for_Meta('/A/B/C','a.TBL','/a/b/c','plotA',10)}
\usage{
Mplot_for_Meta(file_path, file_name, figure_pathy, figure_name, p_col, colour)
}
\arguments{
\item{file_path}{the pathway of meta file}

\item{file_name}{the name of of meta file with suffix}

\item{figure_pathy}{the pathway to save Mplot}

\item{figure_name}{the name of of Mplot}

\item{p_col}{column number of p-value}

\item{colour}{color of Mplot}
}
\description{
Mplot for Meta data generated by METAL
MarkerName : eg. 1:10086:A/C or 1:10086:A:C
Usage eg. Mplot_for_Meta('/A/B/C','a.TBL','/a/b/c','plotA',10)
}
