% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mplot_for_QTLs.R
\name{Mplot_for_QTLs}
\alias{Mplot_for_QTLs}
\title{Mplot for Meta data generated by METAL
MarkerName : eg. 1:10086:A/C or 1:10086:A:C
Usage eg. Mplot_for_Meta('/A/B/C','a.TBL','/a/b/c','plotA',10)}
\usage{
Mplot_for_QTLs(
  file_path,
  file_name,
  qtl_file,
  figure_pathy,
  figure_name,
  p_col,
  snp_col,
  colour
)
}
\arguments{
\item{file_path}{the pathway of meta file}

\item{file_name}{the name of of meta file with suffix}

\item{figure_pathy}{the pathway to save Mplot}

\item{figure_name}{the name of of Mplot}

\item{p_col}{column number of p-value}

\item{snp_col}{column number of MarkerName}

\item{colour}{color of Mplot}
}
\description{
Mplot for Meta data generated by METAL
MarkerName : eg. 1:10086:A/C or 1:10086:A:C
Usage eg. Mplot_for_Meta('/A/B/C','a.TBL','/a/b/c','plotA',10)
}
