% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mplot_for_TWAS.R
\name{Mplot_for_TWAS}
\alias{Mplot_for_TWAS}
\title{Mplot for TWAS results generated by FUSION
Usage eg. Mplot_for_TWAS('/A/B/C','Lung_China_Intron_twasplot_update','TWAS Manhattan Plot of NSCLC','0.1','/A/B/C',11)}
\usage{
Mplot_for_TWAS(
  file_path,
  file_name,
  title,
  Sig_FDR_Thresh,
  outDir,
  color_scheme
)
}
\arguments{
\item{file_path}{the pathway of TWAS result file}

\item{file_name}{the name of of TWAS result file with suffix}

\item{title}{the name and title of of Mplot}

\item{Sig_FDR_Thresh}{Significant FDR threshold of TWAS results}

\item{outDir}{the pathway to save Mplot}

\item{color_scheme}{Color scheme of Mplot including 15 schemes}
}
\description{
Mplot for TWAS results generated by FUSION
Usage eg. Mplot_for_TWAS('/A/B/C','Lung_China_Intron_twasplot_update','TWAS Manhattan Plot of NSCLC','0.1','/A/B/C',11)
}
